+++ pyke/tests/test_kepcotrend.py
@@ -28,3 +28,19 @@ def test_kepcotrend(fitmethod):
     f.close()
     g.close()
     delete("kepcotrend.fits", "log_kepcotrend.txt", False)
+
+def test_kepcotrend_single():
+    kepcotrend(lc, cbv, '1', outfile="kepcotrend.fits",
+               fitmethod="llsq")
+    f = pyfits.open("kepcotrend.fits")
+    g = pyfits.open(
+            get_pkg_data_filename("data/kplr005110407-2009350155506_llc"
+                                  "-kepcotrend-single.fits")
+                   )
+    np.testing.assert_allclose(f[1].data['CBVSAP_FLUX'],
+                               g[1].data['CBVSAP_FLUX'], rtol=1e-4)
+    assert_array_almost_equal(f[1].data['CBVSAP_MODL'],
+                              g[1].data['CBVSAP_MODL'], decimal=5)
+    f.close()
+    g.close()
+    delete("kepcotrend.fits", "log_kepcotrend.txt", False)